refactor(knowledge): extract lib tests to sibling file
primer-knowledge/src/lib.rs was 1718 lines — the largest remaining
single-test-module file in the repo — with a ~739-line
`#[cfg(test)] mod tests` block (lines 978–1718) over the
SqliteKnowledgeBase impl. Move that module verbatim into a new
sibling `src/tests.rs` (crate-root children resolve directly under
src/, unlike non-root modules' `<mod>/tests.rs` form).

lib.rs drops 1718 → 979 lines (−739); it now holds only production
code plus a 2-line `#[cfg(test)] mod tests;` declaration. The only
parent diff addition is `+mod tests;`.

Zero visibility churn: the tests are a descendant module, so
`use super::*` reaches every private parent item unchanged. No
signature changed. Pure test-only relocation — no behaviour change,
no incidental fixes.

Verification: `cargo test -p primer-knowledge --lib` 25 passed (same
as baseline); workspace `cargo test` 51 "test result: ok", 0
failures; clippy + fmt clean.
